Cmssw 7 1 x

+++ GeneratorInterface/SherpaInterface/data/install_hepmc2.sh
@@ -4,18 +4,17 @@
 #  description: BASH script for the installation of the HepMC2 package,
 #               can be used standalone or called from other scripts
 #
-#  author:      Markus Merschmeyer, RWTH Aachen
-#  date:        2009/12/07
-#  version:     2.4
+#  author:      Markus Merschmeyer, RWTH Aachen University
+#  date:        2013/05/23
+#  version:     3.0
 #
 
 print_help() {
     echo "" && \
-    echo "install_hepmc2 version 2.4" && echo && \
+    echo "install_hepmc2 version 3.0" && echo && \
     echo "options: -v  version    define HepMC2 version ( "${HEPMC2VER}" )" && \
     echo "         -d  path       define HepMC2 installation directory" && \
     echo "                         -> ( "${IDIR}" )" && \
-    echo "         -f             require flags for 32-bit compilation ( "${FLAGS}" )" && \
     echo "         -W  location   (web)location of HepMC2 tarball ( "${HEPMC2WEBLOCATION}" )" && \
     echo "         -S  filename   file name of HepMC2 tarball ( "${HEPMC2FILE}" )" && \
     echo "         -C  level      cleaning level of SHERPA installation ( "${LVLCLEAN}" )" && \
@@ -23,6 +22,7 @@ print_help() {
     echo "         -D             debug flag, compile with '-g' option ( "${FLGDEBUG}" )" && \
     echo "         -X             create XML file for tool override in CMSSW ( "${FLGXMLFL}" )" && \
     echo "         -Z             use multiple CPU cores if available ( "${FLGMCORE}" )" && \
+    echo "         -K             keep HEPMC2 source code tree after installation ( "${FLGKEEPT}" )" && \
     echo "         -h             display this help and exit" && echo
 }
 
@@ -34,28 +34,28 @@ HDIR=`pwd`
 # dummy setup (if all options are missing)
 IDIR="/tmp"                # installation directory
 HEPMC2VER="2.03.06"        # HepMC2 version  to be installed
-FLAGS="FALSE"              # apply compiler/'make' flags
 HEPMC2WEBLOCATION=""       # (web)location of HEPMC2 tarball
 HEPMC2FILE=""              # file name of HEPMC2 tarball
 LVLCLEAN=0                 # cleaning level (0-2)
 FLGDEBUG="FALSE"           # debug flag for compilation
 FLGXMLFL="FALSE"           # create XML tool definition file for SCRAM?
+FLGKEEPT="FALSE"           # keep the source code tree?
 FLGMCORE="FALSE"           # use multiple cores for compilation
 
 
 # get & evaluate options
-while getopts :v:d:W:S:C:fDXZh OPT
+while getopts :v:d:W:S:C:DXZKh OPT
 do
   case $OPT in
   v) HEPMC2VER=$OPTARG ;;
   d) IDIR=$OPTARG ;;
-  f) FLAGS=TRUE ;;
   W) HEPMC2WEBLOCATION=$OPTARG ;;
   S) HEPMC2FILE=$OPTARG ;;
   C) LVLCLEAN=$OPTARG ;;
   D) FLGDEBUG=TRUE ;;
   X) FLGXMLFL=TRUE ;;
   Z) FLGMCORE=TRUE ;;
+  K) FLGKEEPT=TRUE ;;
   h) print_help && exit 0 ;;
   \?)
     shift `expr $OPTIND - 1`
@@ -96,12 +96,12 @@ cd ${IDIR}; IDIR=`pwd`
 echo " HepMC2 installation: "
 echo "  -> HepMC2 version: '"${HEPMC2VER}"'"
 echo "  -> installation directory: '"${IDIR}"'"
-echo "  -> flags: '"${FLAGS}"'"
 echo "  -> HepMC2 location: '"${HEPMC2WEBLOCATION}"'"
 echo "  -> HepMC2 file name: '"${HEPMC2FILE}"'"
 echo "  -> cleaning level: '"${LVLCLEAN}"'"
 echo "  -> debugging mode: '"${FLGDEBUG}"'"
 echo "  -> CMSSW override: '"${FLGXMLFL}"'"
+echo "  -> keep sources:   '"${FLGKEEPT}"'"
 echo "  -> use multiple CPU cores: '"${FLGMCORE}"'"
 
 # analyze HEPMC2 version
@@ -127,49 +127,31 @@ export HEPMC2IDIR=${IDIR}"/HEPMC_"${HEPMC2VER}
 
 
 # add compiler & linker flags
-echo "CFLAGS   (old):  "$CFLAGS
-echo "FCFLAGS  (old):  "$FCFLAGS
-echo "FFLAGS   (old):  "$FFLAGS
+echo "CXX      (old):  "$CXX
 echo "CXXFLAGS (old):  "$CXXFLAGS
 echo "LDFLAGS  (old):  "$LDFLAGS
-CF32BIT=""
-if [ "$FLAGS" = "TRUE" ]; then
-  CF32BIT="-m32"
-  export CFLAGS=${CFLAGS}" "${CF32BIT}
-  export FCFLAGS=${FCFLAGS}" "${CF32BIT}
-  export FFLAGS=${FFLAGS}" "${CF32BIT}
-  export CXXFLAGS=${CXXFLAGS}" "${CF32BIT}
-  export LDFLAGS=${LDFLAGS}" "${CF32BIT}
-fi
-CFDEBUG=""
+##MM FIXME
+#  export CXX=""
+#  export CXXFLAGS=""
+#  export LDFLAGS=""
+##MM FIXME
 if [ "$FLGDEBUG" = "TRUE" ]; then
   CFDEBUG="-g"
-  export CFLAGS=${CFLAGS}" "${CFDEBUG}
-  export FCFLAGS=${FCFLAGS}" "${CFDEBUG}
-  export FFLAGS=${FFLAGS}" "${CFDEBUG}
   export CXXFLAGS=${CXXFLAGS}" "${CFDEBUG}
 fi
-echo "CFLAGS   (new):  "$CFLAGS
-echo "FCFLAGS  (new):  "$FCFLAGS
-echo "FFLAGS   (new):  "$FFLAGS
+echo "CXX      (new):  "$CXX
 echo "CXXFLAGS (new):  "$CXXFLAGS
 echo "LDFLAGS  (new):  "$LDFLAGS
 
 # add compiler & linker flags
 COPTS=""
 MOPTS=""
-if [ "$FLAGS" = "TRUE" ]; then
-#    COPTS=${COPTS}" CFLAGS=-m32 FFLAGS=-m32 CXXFLAGS=-m32 LDFLAGS=-m32"
-#    MOPTS=${MOPTS}" CFLAGS=-m32 FFLAGS=-m32 CXXFLAGS=-m32 LDFLAGS=-m32"
-    COPTS=${COPTS}" CFLAGS=-m32 FCFLAGS=-m32 FFLAGS=-m32 CXXFLAGS=-m32 LDFLAGS=-m32"
-    MOPTS=${MOPTS}" CFLAGS=-m32 FCFLAGS=-m32 FFLAGS=-m32 CXXFLAGS=-m32 LDFLAGS=-m32"
-fi
 POPTS=""
 if [ "$FLGMCORE" = "TRUE" ]; then
     nprc=`cat /proc/cpuinfo | grep  -c processor`
-    let nprc=$nprc+1
-    if [ $nprc -gt 2 ]; then
-      echo " <I> multiple CPU cores detected: "$nprc"-1"
+    let nprc=$nprc
+    if [ $nprc -gt 1 ]; then
+      echo " <I> multiple CPU cores detected: "$nprc
       POPTS=" -j"$nprc" "
     fi
 fi
@@ -190,17 +172,19 @@ if [ ! -d ${HEPMC2IDIR} ]; then
     cp ${HEPMC2WEBLOCATION}/${HEPMC2FILE} ./
   fi
   tar -xzf ${HEPMC2FILE}
-  rm ${HEPMC2FILE}
+  if [ ! "$FLGKEEPT" = "TRUE" ]; then
+    rm ${HEPMC2FILE}
+  fi
   cd ${HEPMC2DIR}
   if [ ! -e configure ]; then
     ./bootstrap
   fi
 
-  echo " -> configuring HepMC2 with options "${COPTS}
-  ./configure --prefix=${HEPMC2IDIR} ${momflag} ${lenflag} ${COPTS}
-  echo " -> making HepMC2 with options "${POPTS} ${MOPTS}
-  make ${POPTS} ${MOPTS}
-  echo " -> installing HepMC2 with options "${MOPTS}
+  echo " -> configuring HepMC2 with options "${COPTS} && \
+  ./configure --prefix=${HEPMC2IDIR} ${momflag} ${lenflag} ${COPTS} && \
+  echo " -> making HepMC2 with options "${POPTS} ${MOPTS} && \
+  make ${POPTS} ${MOPTS} && \
+  echo " -> installing HepMC2 with options "${MOPTS} && \
   make install ${MOPTS}
   if [ ${LVLCLEAN} -gt 0 ]; then 
     echo " -> cleaning up HEPMC2 installation, level: "${LVLCLEAN}" ..."
@@ -208,10 +192,15 @@ if [ ! -d ${HEPMC2IDIR} ]; then
       make clean
     fi
   fi
+  cd ${HDIR}
+  if [ "$FLGKEEPT" = "TRUE" ]; then
+    echo "-> keeping source code..."
+  else
+    rm -rf ${HEPMC2DIR}
+  fi
 else
   echo " <W> path exists => using already installed HepMC2"
 fi
-rm -rf ${HEPMC2DIR}
 export HEPMC2DIR=${HEPMC2IDIR}
 cd ${HDIR}
 
@@ -226,7 +215,7 @@ if [ "${FLGXMLFL}" = "TRUE" ]; then
   echo "  <tool name=\"HepMC\" version=\""${HEPMC2VER}"\">" >> ${xmlfile}
   tmppath=`find ${HEPMC2DIR} -type f -name libHepMC.so\*`
   tmpcnt=`echo ${tmppath} | grep -o "/" | grep -c "/"`
-  tmppath=`echo ${tmppath} | cut -f 0-${tmpcnt} -d "/"`
+  tmppath=`echo ${tmppath} | cut -f 1-${tmpcnt} -d "/"`
   for LIB in `cd ${tmppath}; ls *.so | cut -f 1 -d "." | sed -e 's/lib//'; cd ${HDIR}`; do
     echo "    <lib name=\""${LIB}"\"/>" >> ${xmlfile}
   done
